requirements.txt
Flask
flask-cors
python-dotenv
google-genai
google-generativeai
